Bump MySQL 5.7.23 to 5.7.27
Versions between .24 and .27 are affected from
this bug: https://bugs.mysql.com/bug.php?id=95308

MySQL build fails is you do not compile the server. The status of the
bug is "Won't fix" since it is fixed in later versions. As a workaround,
we had to apply a patch file to include a header file. This way the
compilation does not complain.

[#167625725]

Signed-off-by: Sarah Connor <sconnor@pivotal.io>