Do not use VersionId is given version is null
There are cases where a bucket can be versioned and then be switched to
unversioned. Any blob created while the bucket was versioned will always
have a non-null version id. If a user run s3 backuper restorer to a
previously versioned bucket, they will get an error. This is because in
the unversioned code path we always assume that the blob version is
"null", this is not always the case though (as described above). In this
commit we make sure we don't set the version id in the unversioned code
path.

[#167893043]