Commit to the refactor
Previously we had used s3bucket.ForcePathStyleDuringTheRefactor as a
temporary placeholder to describe the value that we pass by default to a
new function parameter introduced in this refactor.

In this commit we realise that we've changed all the behaviour we want
to change, and the remaining uses of ForcePathStyleDuringTheRefactor
just need constants.

[#168837909]

Signed-off-by: Gareth Smith <gsmith@pivotal.io>