Remove duplicate resource and fix jobs semantics
[#181836743]

The two resources don't have the same semantics.
One refers to the current PR being processed while the other
refers to the latest commit of the main branch.

As such, we had to modify some jobs to use a different resource.
For such jobs we reused the semantically identical resource:
`backup-and-restore-sdk-release-main`

The pipeline can already be set without errors, but the logic is
not perfect yet. We will have to create some interdepence among
the existing and the new jobs in the following commits.