Inline bbr sdk test infrastructure (#627)
* Inline bbr-sdk-test-infrastructure pipeline inside the other

[#181836743]

Currently, fly cli seems to completely ignore the new lines since
we have not relocated the resources, resource_types and jobs sections.
We will be relocating these sections in the following commits.

* Move jobs and resources to their corresponding section

[#181836743]

In its current state the pipeline can't be set. It complains about
new jobs not being in a group and some duplicate resources.

* Remove duplicate variables and resource_types

[#181836743]

In its current form the pipeline can't be set.
It complains about the same errors as the previous commit.

* Add new jobs to the shipit group

[#181836743]

In its current form the pipeline can't be set.
It no longer complains about some jobs not being in a group.
It still complains about some resources being duplicated.

* Remove duplicate resource `release`

[#181836743]

The two resources are identical so we don't need to make any
additional modifications to make this change work without errors.

The pipeline is still complaining about another duplicate resource.

* Remove duplicate resource and fix jobs semantics

[#181836743]

The two resources don't have the same semantics.
One refers to the current PR being processed while the other
refers to the latest commit of the main branch.

As such, we had to modify some jobs to use a different resource.
For such jobs we reused the semantically identical resource:
`backup-and-restore-sdk-release-main`

The pipeline can already be set without errors, but the logic is
not perfect yet. We will have to create some interdepence among
the existing and the new jobs in the following commits.

* Create interdependencies between pre-existing jobs and new ones

[#181836743]

We want bbr-sdk to be released automatically every two weeks
only for those changes that have flowed successfully through the new jobs.

For doing that, we changed `passed` conditions for the `check-for-changes` job
to depend on each and all of the new `deploy-` jobs added to the pipeline.