Add job for manually triggering a new release (#664)
We could have achieved a similar result by just manually triggering a new build
for `create-final-patch` job. The downside of that would have been that resource
`every-two-weeks` doesn't get updated. So the pipeline would end up triggering
an automated release in lass than two weeks, as if the manual release had never
happened.

This change adds a new fully manual job named `force-a-new-release` which simply
updates the `two-weeks-resource` creating a new version immediately with the
current date and time. This effectively triggers the creation of a new release
and updates the `every-two-weeks` scheduler to not create an automated release
until two weeks have passed.