use mysql-boost source tarball for 8.0
This commit just updates the autobump-mysql.sh script so that it will
fetch and verify mysql-boost-8.0.*.tar.gz source tarballs. This avoids
needing to maintain an additional MySQL 8.0 boost dependency, as the
require boost headers are included in the source tarball itself.

MySQL 5.7 (and earlier) are left alone.  While Oracle provides
mysql-boost-5.7.* tarballs, there are a few build script bugs that cause
compilation to fail.