Feature/mysql 8 support (#702)
* mysql: support tls in mysql system_tests when using pxc-release

- Add mtls configuration to ci manifest
- Adjust dns resolution to support bosh dns

Authored-by: Andrew Garner <garnera@vmware.com>

* Increase default eventually timeout in integration tests

Observed timeouts locally.  Tests pass when timeout is increased.

Authored-by: Andrew Garner <garnera@vmware.com>

* database-backup-restore: add mysql-8.0 integration test

This is a failing test.  Did some minimal test setup to get some vaguely
useful error messages.

Internally this utility defaults to using mysql 5.7 client to discover
the mysql version. If this is not configured properly the utility fails
with essentially no error output.

Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Ryan Wittrup <rwittrup@vmware.com>

* support mysql-8.0 in database-backup-restore

* add database-backup-restorer-mysql-8.0 package

- Missing mysql-boost-8.0.*.tar.gz blob in blobstore, because I lack
  access to the release blobstore.

* add ci/manifests/ops-files/dev-release.yml

This helps deploy locally to create a one-off backup-and-restore-sdk
release.

This may be a throw-away commit, but I found it useful when iterating
locally.

* use mysql-boost source tarball for 8.0

This commit just updates the autobump-mysql.sh script so that it will
fetch and verify mysql-boost-8.0.*.tar.gz source tarballs. This avoids
needing to maintain an additional MySQL 8.0 boost dependency, as the
require boost headers are included in the source tarball itself.

MySQL 5.7 (and earlier) are left alone.  While Oracle provides
mysql-boost-5.7.* tarballs, there are a few build script bugs that cause
compilation to fail.

* mysql_test: Fail if mysql 8 and xenial

- As of today mysql8 cannot be compiled against xenial stemcells.
- This commit will present the user with a useful error message if they
  are trying to backup or restore a mysql 8 database when bbr is
  compiled against xenial

[#182920153](https://www.pivotaltracker.com/story/show/182920153)

Co-authored-by: Kyle Ong <kyleo@vmware.com>
Co-authored-by: Andrew Garner <garnera@vmware.com>

* packaging/mysql-8.0: Exit 0 if xenial

mysql-8.0 does not support gcc 7.1 or clang 5.

[#182954679](https://www.pivotaltracker.com/story/show/182954679)

Co-authored-by: Kyle Ong <kyleo@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>

* Add mysql8 blob to blobstore.

[#183155304]

* Fix mysql8 unit tests

[#183155304]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

* Revert "mysql: support tls in mysql system_tests when using pxc-release"

This reverts commit 0cb422f537b21665b347d377534538a33b51653a.

Signed-off-by: Gareth Smith <sgareth@vmware.com>
Co-authored-by: Andrew Garner <garnera@vmware.com>
Co-authored-by: Ryan Wittrup <rwittrup@vmware.com>
Co-authored-by: Kyle Ong <kyleo@vmware.com>
Co-authored-by: Kim Bassett <kbassett@vmware.com>