Add database-unit-tests to GHActions workflows
[#183164582]

Notice that we are migrating an existing Concourse task at
/ci/tasks/database-unit-tests. Also notice that we removed
every `ssh` reference in that task, including github-sdk-key
input. We are pretty sure we don't need this anymore and it
was simply required in the past because we were using `glide`
as a package manager. See this commit for more info:
https://github.com/cloudfoundry/backup-and-restore-sdk-release/commit/b6b5685fbffacf5700c5fcc1b814f0f8c99fb979#diff-e8a87dc65a0a105ed236b6898d78ace032435af2201bac35dc3a21d63789f46eR14

Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>