Use docker-compose to simplify testing offline and in GHActions
[#183164582]

This works because relative paths inside a docker-compose.yml are
resolved as relative to the docker-compose.yml itself instead of
to the current working folder of the shell.

Notice that we are using `docker compose run` instead of
`docker compose up` that's because `docker compose up` doesn't
return the same **exit code** as the process running inside the
container by default. Conversely, `docker compose run` does.