Add a unit tests workflow for sdk-release-tests (#714)
* Add a unit tests workflow for sdk-release-tests

[#183164582]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

* Add s3-blobstore unit tests

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>

* Add set-safe-directory option to repo

- That's because CI complained with the following error `Use
  -buildvcs=false to disable VCS stamping`.
- The test is trying to query some information from git but the way it
  gets cloned in GH Actions does not allow it to.

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>

* Revert "Add set-safe-directory option to repo"

It did not work, possibly because we are using submodules. Also the
setting was set to `true` by default.

This reverts commit 0d8bbffc32c098398a23d7b3c1b7c8a9016128bc.

* Add `-buildvcs=false` to S3 Blobstore test

- That's so we can run `go build` within the tests without worring how
  GH actions clones repos.

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>

* Use golang image instead of backup-and-restore one

Hopefully this should be lighter weight, and easier to keep up to
date.

[#183164582]

* Rename workflow

- Now we are running many jobs in the same workflow but all of them are
  Unit tests

* Add Azure Blobstore unit tests

* Add GCS blobstore unit tests

* Pin golang docker image to current latest

[#183164582]

Also:
- Remove unneded ENV commands
- Stylistic refactor of whitespaces

Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>

* Add database-unit-tests to GHActions workflows

[#183164582]

Notice that we are migrating an existing Concourse task at
/ci/tasks/database-unit-tests. Also notice that we removed
every `ssh` reference in that task, including github-sdk-key
input. We are pretty sure we don't need this anymore and it
was simply required in the past because we were using `glide`
as a package manager. See this commit for more info:
https://github.com/cloudfoundry/backup-and-restore-sdk-release/commit/b6b5685fbffacf5700c5fcc1b814f0f8c99fb979#diff-e8a87dc65a0a105ed236b6898d78ace032435af2201bac35dc3a21d63789f46eR14

Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>

* Introduce a generic Dockerfile to run golang tests

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>
Signed-off-by: Iain Findlay <fiain@vmware.com>

* Refactor database tests to use generic Dockerfile

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>
Signed-off-by: Iain Findlay <fiain@vmware.com>

* Refactor GCS tests to use generic Dockerfile

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>

* Refactor S3 tests to use the generic Dockerfile

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>
Signed-off-by: Iain Findlay <fiain@vmware.com>

* Trigger GH actions

Signed-off-by: Diego Lemos <dlemos@vmware.com>

* Refactor Template tests to use generic Dockerfile

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>
Signed-off-by: Iain Findlay <fiain@vmware.com>

* Put GH Action jobs in alphabetical order

[#183164582]

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>
Signed-off-by: Iain Findlay <fiain@vmware.com>

* Use docker-compose to simplify testing offline and in GHActions

[#183164582]

This works because relative paths inside a docker-compose.yml are
resolved as relative to the docker-compose.yml itself instead of
to the current working folder of the shell.

Notice that we are using `docker compose run` instead of
`docker compose up` that's because `docker compose up` doesn't
return the same **exit code** as the process running inside the
container by default. Conversely, `docker compose run` does.

* Make docker compose work with older versions

Older versions of docker-compose can't detect which version of the
docker-compose.yml file format we're using, so we need to specify it
explicitly.

Signed-off-by: Gareth Smith <sgareth@vmware.com>

[#183164582]

* Prepare for adding non-unit tests to docker-compose file

Previously ever workflow in the docker compose file was for running
unit tests. Now we begin those workflow names with "unit-". This means
that if we add non-unit test workflows (for example, MySQL tests), and
sort the workflows alphabetically, they will end up in sensible
logical groupings.

[#183164582]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

* Add human interface to unit tests

Now that we're happy with our docker-driven unit tests, we make them
easy to run for any dev who has cloned the repo.

Previously `make unit` assumed you had the correct versions of a large
number of dev tools installed. You can still use this flow with `make
local-unit`, but by default we use the docker tests.

[#183164582]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

Signed-off-by: Gareth Smith <sgareth@vmware.com>
Signed-off-by: Diego Lemos <dlemos@vmware.com>
Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>
Signed-off-by: Iain Findlay <fiain@vmware.com>
Co-authored-by: Fernando Naranjo <fnaranjo@vmware.com>
Co-authored-by: Diego Lemos <diego@diegolemos.net>
Co-authored-by: Diego Lemos <dlemos@vmware.com>