Add airgap tests (#704)
* Add airgap tests

Copy the testing pattern from the bosh-addons pipelines. Claim a
CF-deployment environment. Cut the internet. Upload our bosh release,
and see if it compiles.

This won't check that everything properly functions behind an airgap,
but it will check that we can at least compile.

[#183164582]

* Make pipeline work even when lots of PRs come in at once

Enable the autosync_pr feature of the PR queue resource. This makes
the resource automatically merge main into the current PR before we
start the pipeline. That way, when we're done testing, we're pretty
confident that it should be safe to merge.

[#183164582]

Signed-off-by: Gareth Smith <sgareth@vmware.com>

Signed-off-by: Gareth Smith <sgareth@vmware.com>
Co-authored-by: Fernando Naranjo <fnaranjo@vmware.com>
Co-authored-by: Cryogenics CI Bot <mapbu-cryogenics@groups.vmware.com>