Remove from matrix failing unsupported versions of MYSQL
[#183164582]

There are chances that the versions removed could also work
with our current release but they need some packages not included by default in
the Dockerimage. Doing the extra effort to make the pass is irrelevant to the
BOSH release itself, since the filesystem is not identical to what will be included
in the VM.

We might need to investigate how to reduce this filesystem mismatch if possible.

Signed-off-by: Fernando Naranjo <fnaranjo@vmware.com>