Improve debugging experience by mounting the whole repo instead of a subfolder
[#183164582]

By doing this any error message in the logs would point to the exact path where
the problem occurred instead of a docker-mounted path in an arbitrary location.

This change pursues reducing the knowledge burden introduced by the use of Docker
by trying to make the Docker integration as thin and transparent as possible.
Ideally, not requiring any Docker knowledge to be able to work with the codebase.