Reduce certs handling in `Dockerfile` and `docker-compose.yml`
[#183164582]

By aglutinating the three previous variables each one pointing to
a different certificate into a single variable pointing to a folder
we should contain the three files.
These files are expected to always be named the same so there isn't
a need to specify the full path for each of them individually.