Test compilation against diff stemcells and improve system tests
[#183164582]

To test compilation against diff stemcells we use multistage Docker containers
to download bosh-lite warden tarballs and using the rootfs included in the tarball
to create a container FROM scratch.

Inside that dockerized stemcell we run the release packaging scripts which allows
testing compilation against different stemcell using GHActions matrix strategies.

Now that we are able to compile the release directly in GHActions, we can improve
system tests by leveraging the binaries built on-the-fly from the blobs.
This approach has huge benefits compared to the previous one:
- Tests are more true to reality
- Bosh blobs are fully tested (instead of testing just the GO code)
- Compilation is being tested against different stemcells