Only run PRs marked as "ActivePR" using a label
Before this change every PR workflow got triggered immediately on every new PR causing a lot of overhead on GitHub-Hosted runners.

Morever, since we want any PR we merge to be up-to-date with its base branch, it makes no sense to test all PRs simultaneously since we are only able to merge 1 PR per base branch at a time.

This change modifies existing workflows to only run when a new label is added to the PR.

Also, a new Workflow was created to be in charge of adding these labels to the opened PRs in due time.