Support postgres 15 (#785)
* Support postgres 15

[#183414909]

* Fix unit-tests for postgres 15

[#183414909]

* Stop testing old unsupported versions of postgres

[#183414909]

* Fix issue with BOSH not compiling postgres 15 when deploying

[#183414909]

Adding postgres 15 as a dependency for database-backup-restorer job
was missing.

* Upload release when starting the BOSH-in-Docker Director

[#183414909]

This change is not just for the sake of efficiency. Previous approach had issues:
- `upload-release` (from source) ignores local changes making local debugging imposible
- `create-release --force --tarball` creates new files and folder in the repo which fail
  if we try to run create-release more than once

https://github.com/cloudfoundry/bosh-cli/issues/604

The best solution I could think of was ensuring that we upload the release only once.
We can iterate on this design later and make it cleaner or more robust.

* Ensure make reports a non-zero error code when any test fails

[#183414909]

Because of the way we are executing bash commands inside the Makefile
(multiple commands get concatenated as a multiline single instruction)
Make was incorrectly reporting success even when some tests failed.

* Copy the repo into the Docker images instead of mounting

[#183414909]

This increases encapsulation and isolation because:
- Tests can write to the filesystem (accidentally or intentionally)
  and still those changes won't affect other tests since the changes
  will dissappear with the container
- It is easier to reason about the tests without having to worry
  about mutability and side-effects to the filesystem such as
  `create-release` adding new folders to the repo
- It is easier to debug issues and ensure the tests are doing what
  they are expected since they can't introduce unintended changes
  or leave traces that affect subsequent tests

However, with this change we are forced to use `docker-compose up`
instead of `docker-compose run` because with `up` we can enforce
that a new container image is built on every run using `--build` flag
(Docker CACHING takes care of performing this step very efficiently)

Without `docker-compos up --build` it is harder to ensure that tests
are running against latests changes in the repository (it might reuse
an older image) making local debugging and making local changes a pain.

* Fix typo in docker service name

[#183414909]

* Replace deprecated mariadb version in Azure

[#183943473]

* Fix postgres clientcert incompatible behaviour across versions

[#183414909]

Co-authored-by: Cryogenics CI Bot <mapbu-cryogenics@groups.vmware.com>