Simplify gh workflow (#811)
* Change workflow triggering conditions

[#184013456]

Simplify the workflow to not depend on any labels.
Also, we don't want to run any tests on PR creation,
instead run only on new commits or when
the PR gets assigned to some user.

This prevents saturating GH runners after Dependabot
runs, when a lot of PRs get created simultaneously

* Assign PR to token owner during check

[#184013456]