Download gnu-keyring on every execution
[#184179659]

We can't use SCRIPT_RELATIVE_DIR as a
way to reference the current directory because
autobump-ncurses.sh gets `eval`ed.
So the variable will point to the calling script instead.

I believe the initial implementation didn't
suffer this problem because the core script
was colocated alongside the product scripts,
but this is no longer true.

The core script is now store in a different repo:
https://github.com/pivotal/cryogenics-concourse-tasks/tree/main/deps-automation/bump-package