chore: fix bumping steps
The default image for the bump task has changed to be the cryogenics
essentials image, which makes sense for other products. But for this
product, the bump script requires `xmllint` which we ideally don't want
to put in the cryogenics-essentials image as it's meant to be a lean
image with just the essentials. So for that tasks that need `xmllint` we
will override the default and restore the previous behavior of using the
pcfplatformrecovery/backup-and-restore image for the bump tasks.

[#185859706]