vault resource for director is deprecated
this is happening as a flyby to avoid a lot of manual work because
of an ci change. The setup-infrastructures pipeline now output a
director source file that can be consumed. The main reason is to
remove the vault resource usage to be able to move this pipeline to
runway where as of 29.08.23 no service account access to write secrets
to vault exists