Update pipeline (#1204)
* add autobumper for openssl-1 blob

[#183639201]

this adds an autobumper for the openssl-1 package of the release

* add job to serial group

[ #183639201 ]

* vault resource for director is deprecated

this is happening as a flyby to avoid a lot of manual work because
of an ci change. The setup-infrastructures pipeline now output a
director source file that can be consumed. The main reason is to
remove the vault resource usage to be able to move this pipeline to
runway where as of 29.08.23 no service account access to write secrets
to vault exists

* dont use bosh for read values from yaml

if the environment has BOSH_ ENV vars set with bad values (in this
tasks case BOSH_CA_CERT="" is set as a default), then the bosh-cli
will refuse to work with an error like:

```
invalid argument for flag `--ca-cert' (expected opts.CACertArg): Expected CA cert to be non-empty
```

avoid this by using yq.