Use cfd airgapped shepherd (#1325)
* remove need to create custom firewall rule

[ #186163798 ]

we want to switch this CI to use shepherd environments. These do not
expose the gcp service account key and we cannot create custom security
group rules.

This PR works around this by utilizing the vm_extension present in cfd that
is normally used for tcp routers. This vm_extension will apply a fw rule that
opens all non privileged ports on the VM including the required 3306 and 5432
database default ports.

* Move BBR-SDK to shepherd [#186163798]

---------

Co-authored-by: nouseforaname <34882943+nouseforaname@users.noreply.github.com>
Co-authored-by: Cryogenics CI Bot <mapbu-cryogenics@groups.vmware.com>