Fix version pagination by adding default batch size
**Issue:**

We realised that whilst bumping `aws-sdk-go-v2/services/s3` to version
`1.44.0` we found a breaking change where a default value for `MaxKeys`
in the `ListObjectVersionsInput` struct was being set to `0`, resulting
in 0 versions being returned. It should have been `1 000`.

```
input := &s3.ListObjectVersionsInput{
	Bucket:  aws.String("tf-acc-test-7828939106372858870"),
	MaxKeys: aws.Int32(1000),
}
```

More details: https://github.com/aws/aws-sdk-go-v2/issues/2379

**Workaround:**

We are explicitly setting the `MaxKeys` to `1_000` as it should be.

[#186503787]

Signed-off-by: Diego Lemos <dlemos@vmware.com>