switch mysql bumper, delete obsolete tasks / resources
[#186663246]

replace old mysql bumper script with generalized pattern.

it has served us well but it is time to move on. In an effort
to generalize how cryogenics maintains their products, we're
moving away from customer bumper scripts for single packages
to a generalized model of concourse resources and shared tasks.

- enable force on push for bumps

if the bumper runs a 2nd time,  before the previously created PR got merged
it will fail without the force flag. To avoid this, enable force pushing the
same commit (that already got created by the first run of the bumper for a given
version X)

- remove daily trigger

when the resource daily triggers, we potentially already created a PR and in that
case CI will break. It should only trigger on actual new versions of mysql which
we now can detect because we switched to the ftp / http resource to acquire the
blob.