drop obsolete dbs
[#186663246]

mysql 5.6, 5.7 and postgres-10 are out of support.

- cleanup the release removing unneeded packages / blobs
- clean ci from those versions
- clean Makefile testmatrix

- update the version check binary to the lowest in support major
versions of postgres and mysql

- mysql tests needed updating of their setups. The lowest
mysql binary version available in the bbr sdk is used to
run the SQL statement that checks the version of the targeted
server. That means that the invocation length incremented by
1 since mysql80 binary is now used to run the version check
after we removed support for 5.7. Similarly this updates the
expectations for the args for the restore / backup commands
to now use the 2nd Invocation to check against.