Maintenance/clean out of support db version (#1455)
* switch mysql bumper, delete obsolete tasks / resources

[#186663246]

replace old mysql bumper script with generalized pattern.

it has served us well but it is time to move on. In an effort
to generalize how cryogenics maintains their products, we're
moving away from customer bumper scripts for single packages
to a generalized model of concourse resources and shared tasks.

- enable force on push for bumps

if the bumper runs a 2nd time,  before the previously created PR got merged
it will fail without the force flag. To avoid this, enable force pushing the
same commit (that already got created by the first run of the bumper for a given
version X)

- remove daily trigger

when the resource daily triggers, we potentially already created a PR and in that
case CI will break. It should only trigger on actual new versions of mysql which
we now can detect because we switched to the ftp / http resource to acquire the
blob.

* drop obsolete dbs

[#186663246]

mysql 5.6, 5.7 and postgres-10 are out of support.

- cleanup the release removing unneeded packages / blobs
- clean ci from those versions
- clean Makefile testmatrix

- update the version check binary to the lowest in support major
versions of postgres and mysql

- mysql tests needed updating of their setups. The lowest
mysql binary version available in the bbr sdk is used to
run the SQL statement that checks the version of the targeted
server. That means that the invocation length incremented by
1 since mysql80 binary is now used to run the version check
after we removed support for 5.7. Similarly this updates the
expectations for the args for the restore / backup commands
to now use the 2nd Invocation to check against.

* we do not know what this passed constraint solves.

we do know that it currently breaks ci without because it
forces us to restart a PR test at a previous stage to get the
newer version of cryo tasks and we don't like it.

* backfill pg-15 tests

enable running tests against PG-15 as desired in [#186845191]

- do not run pg-15 no tls tests against rds

  the default parameter group for postgres 15 rds instances sets
  `rds.enforce_ssl` to 1, which means enabled.
  see:
  https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/PostgreSQL.Concepts.General.SSL.html

  instead of updating the test db to be insecure, drop non tls tests.

* add test ensuring that we do not accept the oos 5.7 version

* enable removing multiple blobs
[#186859120]
fix ci:

```
+ NEW_BLOBNAME=mysql-((redacted))-linux-glibc2.17-x86_64-minimal.tar.xz
+ echo 'removing old blob(s) mysql/mysql-5.6.51.tar.gz mysql/mysql-5.7.44.tar.gz mysql/mysql-8.0.35-linux-glibc2.17-x86_64-minimal.tar.xz'
removing old blob(s) mysql/mysql-5.6.51.tar.gz mysql/mysql-5.7.44.tar.gz mysql/mysql-8.0.35-linux-glibc2.17-x86_64-minimal.tar.xz
++ tr -dc ' '
```

which is fine since we're dropping support for oos mysql versions.

* fix connectionString for sslkey and sslcert

commas are not the right separator for the elements in url params.

* fix deployed mysql version

the current ci setup will deploy pxc release with a mysql version of 5

caused by yaml rendering and resource config issuies.

- rendering issue:

  the value for mysql_version in the pipeline is `8.0`
  the value for mysql_version in the rendered manifest deployed on bosh is `8`.

  since yaml supports float and the multiple yaml parsers at play
  are trying to be smart (concourse, bosh, bosh-cli) we end up not
  properly propagating the exact required value.

- pxc-release resource config

  Fixing the above wouldn't help the case because we deploy the
  pxc-release at version 0.50.0, instead of latest. And that version does not
  yet provide the version picking mechanism.

  Unfortunately the release config for pxc pointed to:
  `https://bosh.io/d/github.com/cloudfoundry-incubator/pxc-release`.

  Note the `incubator` contained in the url

  This repo has been moved to `cloudfoundry/pxc-release`
  so we're getting the state from before the move. I'm not sure if this is
  an error on bosh.io end or github or ???. What I do know is that calling the
  configured URL will start a download for `v0.50.0`.

  To remedy this, quote all usages of `mysql_version` so that it is always
  handled as a string by yaml. Additionally update the bosh release config
  so it will start downloading the latest available release.

---------

Co-authored-by: Gareth Smith <sgareth@vmware.com>
Co-authored-by: Cryogenics CI Bot <mapbu-cryogenics@groups.vmware.com>