add cmake to enable compililing mariadb w. ossl3
[#186657448]

when we call cmake with `-DWITH_SSL` in the packaging script for
mariadb we will use cmake functionality that will look into the
specified DIR to

- find locations for specific .so files
- find the specific ossl version found in $DIR

this functionality will not work on cmake 3.5 present on xenial stemcells.
it will fail parsing the version and output `.0.0` making the entire configure
fail. to enable linking against ossl3 on xenial, conditionally compile cmake
so it can be used when compiling mariadb