Migrate to new azure sdk (#1555)
* First commit of Azure SDK re-work, needs refactoring.

[#187088873]

Signed-off-by: Iain Findlay <fiain@vmware.com>

* Extract buildServiceClient in container.go

[#187088873]

* Rename URL variable for clarity

[#187088873]

* Extract permissions expiry for clarity

[#187088873]

* Add a utility for sanitising containerURLs

[#187088873]

* Sanitise all URLs before potentially logging them

Previously we returned URLs raw in some of our error messages. These URLs could
contain secrets (like SAS keys[1]), and could also end up in logs. To avoid
this, we want to sanitise all query parameters in the URLs we put into error
messages. We explicitly allow "snapshot" queries, since we think those are both
safe, and also particularly useful for debugging.

[1] https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/container#Client.GetSASURL

[#187088873]

* Give URLs in error messages instead of just names

Previously we might log that we failed while copying foo to foo. Now we will
say we failed while copying from https://place1/foo to https://place2.foo. This
will hopefully be more useful.

[#187088873]

* Remove last vestigest of old azure library

It's no longer possible or desirable to do the kinds of network-retries that we
did with the old library. So we've removed a bunch of test logic an infra that
supported those retries.

[#187088873]

---------

Signed-off-by: Iain Findlay <fiain@vmware.com>
Co-authored-by: Iain Findlay <fiain@vmware.com>
Co-authored-by: Cryogenics CI Bot <mapbu-cryogenics@groups.vmware.com>